requirements.txt
discord.py
groq
python-dotenv
google-generativeai
